FMT: dont remove mandatory commas & clean tests up
closes #1566
