--- src/test/resources/org/rust/ide/formatter/fixtures/issue543a.rs
@@ -1,1 +1,0 @@
-pub type TeraResult<T> = Result<T, TeraError>;
