--- src/test/resources/org/rust/ide/formatter/fixtures/issue543b_after.rs
@@ -1,6 +1,0 @@
-impl_rdp! {
-    grammar! {
-        expression = _{ paren ~ expression? }
-        paren      =  { ["("] ~ expression? ~ [")"] }
-    }
-}
