--- src/test/resources/org/rust/ide/formatter/fixtures/issue543c_after.rs
@@ -1,8 +1,0 @@
-fn main() {
-    if previous_end < token.start {
-        space_tokens.push((
-            insert_at,
-            Token::new(Rule::text, previous_end, token.start)
-        ));
-    }
-}
