--- src/test/resources/org/rust/ide/formatter/fixtures/line_breaks/blocks.rs
@@ -1,4 +1,0 @@
-fn main() {
-    let foo = { foo(123, 456,
-                    789) };
-}
