--- src/test/resources/org/rust/ide/formatter/fixtures/line_breaks/blocks2_after.rs
@@ -1,3 +1,0 @@
-fn main() {
-    let foo = { foo(123, 456, 789) };
-}
